POC: Codesearch result quality benchmarks (#9780)
This program uses [discounted cumulative
gain](https://en.wikipedia.org/wiki/Discounted_cumulative_gain) (along
with some strategies outlined in the
[limitations](https://en.wikipedia.org/wiki/Discounted_cumulative_gain#Limitations)
section) to score actual codesearch results against sets of "ideal"
results.

The ideal results are hand curated, and expected to be from a fixed set
of documents. In this POC, the buildbuddy repo is indexed at a
particular revision, queries are run against it, and the actual results
are scored against the ideal results.

The "ideal results" proto has an entry for matched line numbers, but
they aren't currently used in scoring the results. The initial idea was
that this dataset could also be used in a set of correctness regression
tests, where we ignore ranking but ensure we get back exactly the
expected results.

There are open questions around the best way to use this - I think a
nice state would be that this runs automatically when codesearch
dependencies change, and the results are reported to PR, so that any
regression in scores could be investigated. Maybe a workflow?

This would be a useful tool to get up and running before making any
significant attempts to improve result ranking, so that progress could
be measured.