Add Details methods to FlagProvider (#9805)
It's sometimes necessary to know which `variant` was selected, e.g.
https://github.com/buildbuddy-io/buildbuddy/pull/9799

This PR adds "Details" versions of each FlagProvider method so that this
info can be retrieved. There are some other things in the flagd Details
which might be interesting (such as Error) so I went with an interface
exposing a `Variant()` method which we can extend later, rather than
just exposing a `variant` string.