Allow fetching ExecuteResponse from tools/cas (#9811)
Allow passing an execution ID to fetch the corresponding ExecuteResponse
proto that was cached for that execution. This is useful for some
scripting around `replay_action` that I'm doing (given an `execution_id`
I need to know which `executor_host_id` executed the action).

Also rename `--digest` to `--resource` since at this point we're already
parsing it as a resource name.

(At some point we've gotta rename this to `tools/cache` btw :P)