Don't block on mutexes when calling atime_updater.Enqueue functions (#9809)
This PR restructures the atime_updater to use a channel for enqueuing
updates, instead of doing it synchronously. Updates are formed and sent
to the channel only if sending to the channel will not block. Then,
there is a goroutine that consumes updates from this channel and batches
them into the per-group, and per-instance-name-and-digest-function maps.
There is still another goroutine that consumes these batches and sends
them to the backend, as before. This means that calls to atime_updater
Enqueue functions should now never block, which should help cache proxy
throughput under load a bit. This makes the code a little bit more
complicated, but I think it's a reasonable tradeoff to make.

I also changed the tests to call the lifecycle functions on
atime_updater directly, instead of relying on a fake clock and sleeps,
which should make it a bit faster and less flaky.

Related Issues:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4944