trace_viewer: Adjust auto-zoom for filtered spans
Make sure that we also trigger the zoom adjustment when a span is in view,
but is too small for the canvas to render.

Dial down the MIN_SPAN_PIXEL_WIDTH to 1px so that we would only adjust
the zoom when it is absolutely needed to view the span. This should
reduce the cases where the zoom adjustment trigger, which can be
disorientating for some users.