Reduce errors from clickhouse replication lag when fetching executions (#9833)
Executions data is stored in Redis before it's finalized, and eventually
finalized and stored in Clickhouse. When written to Clickhouse, the
Redis entries are deleted. Due to replication lag with clickhouse
ReplicatedMergeTree tables, there can be a window when the data is
unreadable - it's been deleted from Redis, and the Clickhouse read may
go to a replica that doesn't have the data yet

This PR deletes the data from Redis with some buffer time instead, by
setting a 5min TTL on the entries. In the last 2 weeks, [the max
replication lag was
~2min](https://metrics.buildbuddy.io/d/clickhouse-operator/clickhouse-operator-dashboard?orgId=1&from=2025-06-19T05:12:30.621Z&to=2025-07-07T21:43:27.293Z&timezone=browser&var-exported_namespace=$__all&var-chi=$__all&var-hostname=$__all&refresh=1m&viewPanel=panel-7)

More context
[here](https://buildbuddy-corp.slack.com/archives/C0495TM9UUE/p1751914394107319)