Improve logging and metrics in the migration cache (#9842)
- Use log.Ctx... methods so it's easier to spot failures just for one
group ID
- For read methods, only start a destination read if the source
succeeded. This means we only log misses when we expect hits and we only
increment hit and miss metrics when we expect a hit. This makes it much
easier to tell what percentage or requests were hit in the source but
missing in the destination. Finally, this means that we don't enqueue
copies when the source doesn't contain the resource.