Add a concurrency limit on FileWriter ops.
FileWriters are used to write artifacts to disk by the cache
implementations. When there are bursts of file writer operations, it can
cause the Go runtime to create an unbounded number of system threads as
file writer ops end up blocked on syscalls.

The concurrency limit places an upper bound on the system threads
created via this code path.