Test add replicas from different intermediate states (#9863)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4913

Also, fix a bug I found around waiting for replica to catch up. This
call can
only be called on the machine with the new shard and calling it at the
promoteToVoter doesn't have any effect on the new shard.

Instead, I make sure we wait for the new shard to catch up when the
shard
already exists.