Add offset and limit support for GCS readers
In practice, this never happens, because things we write to GCS are always compressed, and when reading compressed data with an offset and/or limit, [pebble_cache.reader](https://github.com/buildbuddy-io/buildbuddy/blob/7dd9ee4ccba0e92eac1b1c37d7406241f7a26827/enterprise/server/backends/pebble_cache/pebble_cache.go#L3300-L3303) uses `0, 0` for the underlying reader, and only applies the limit to the wrapped reader.

That said, I thought it was simpler to fix this than to add preconditions that ensure offset and limt are 0.