Increase test size for  migration_cache_test (#9904)
I guess when I ran it last, we weren't fully loaded so a 60s timeout was
long enough, but when we're under load, we need a bit more time.

I also decreased the number of parallel requests in
TestCopyDataInBackground. With only 2-8 CPUs for the test action, we're
not getting anything out of running 1000 parallel requests.