networking_test: Update ping target to google.com
Currently example.com is resolved to these IPv4 addresses
```
> dig example.com 8.8.8.8 | grep 'example.com'
; <<>> DiG 9.10.6 <<>> example.com 8.8.8.8
;example.com.			IN	A
example.com.		277	IN	A	23.192.228.80
example.com.		277	IN	A	23.215.0.136
example.com.		277	IN	A	23.215.0.138
example.com.		277	IN	A	96.7.128.175
example.com.		277	IN	A	96.7.128.198
example.com.		277	IN	A	23.192.228.84
```

Out of which, the ones in `96.*.*.*` range are not pingable

```
> ping -c1 -W2 96.7.128.175
PING 96.7.128.175 (96.7.128.175): 56 data bytes
--- 96.7.128.175 ping statistics ---
1 packets transmitted, 0 packets received, 100.0% packet loss
exit 2

> ping -c1 -W2 96.7.128.198
PING 96.7.128.198 (96.7.128.198): 56 data bytes
--- 96.7.128.198 ping statistics ---
1 packets transmitted, 0 packets received, 100.0% packet loss
exit 2
```

Change our tests to use google.com instead.