[raft] start the zombie janitor after the cluster is configured (#9960)
I plan to remove zombieMinDuration as a heuristic to detect zombies.
Instead,
I want to check against ranges in txnRecords in meta range. However,
when we
start the clusters, we didn't use transactions to update the range
descriptors
and also we don't need txn to update the meta range. By starting the
janitor
after the cluster is configured, we can prevent the zombie janitor from
deleting
shards that are being created during this phase.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/5402