Allow using gRPC for the GCS client (#9937)
After a bunch more experimenting, I think this actually does reduce the
minimum GCS read latency from ~40ms to ~30ms. It's hard to know for sure
without sending somewhat real data, so put this behind a flag that we
can enable in dev first.

Also, I realized that due to a somewhat [recent
change](https://github.com/googleapis/google-cloud-go/pull/11878) in the
GCS library, DirectPath now works even if we provide credentials. So
that means that as soon as we enable gRPC, we will also use DirectPath
when running on GCE.