podman: fix race on cidfile (#9984)
In #9757, we reduced the cgroup stats polling rate by increasing the
ticker interval from 50ms to 500ms. This change uncovered a bug where
the execution code could run and exit very quickly, removing the
container id(CID) file as a result. A slower stat collection rate might
continue to run after the file has been removed, causing it to wait for
the entire duration of pollCIDTimeout, which is currently 15 seconds.

Remove the `--rm` flag in podman run args in favor of explicit cleanup
through c.Remove().
Switch `c.Remove()` to use `podman rm -f` so that we also cleanup the
container resources (file system) after sending SIGKILL to the process.

Add a warn log to the final stat collection run to ensure no issues are
missed.