[SS] Support multiple remote snapshot read policies (#9795)
This PR adds a platform property to prioritize local snapshots if
available, to reduce remote cache download

---
Currently this could happen:

Workflow on pr1 goes to executor1. It pulls the master snapshot from the
remote cache and saves a snapshot to the local cache.

A master workflow on another executor updates the master snapshot in the
remote cache.

Workflow on pr2 goes to executor1. It will pull the newest master
snapshot from the remote cache. The local snapshot chunks that pr1 saved
locally may not be usable, so all chunks must be pulled from the remote
cache

--- 
In order for this PR to be effective, I think we'd also need something
like this [PR](https://github.com/buildbuddy-io/buildbuddy/pull/9851) so
that local snapshots will actually be usable. I think we may also need
to enable `workflow_default_branch_routing_enabled`, which directs the
task router to try and route workflows with the same default branch to
the same executor. Otherwise in the example above, pr1 and pr2 could
both go to random executors, and wouldn't be able to share local
snapshots