Add support for `bb --version` (#9996)
Addresses buildbuddy-io/buildbuddy-internal#5072.

This looks like a hack, but it is actually handling a special case.
`--version` is not a startup option, it is an argument that, when passed
as the sole argument to bazel, shortcuts all other processing and does
not start a bazel server to allow the version to be checked without that
startup cost.