Frontend: Handle ANSI SGR codes that do not reset at end of line (#10016)
This fixes a bug where we were implicitly resetting ANSI SGR state at
the end of every line. This behavior is not correct; ANSI SGR state
should only reset if we encounter an ANSI reset code. To acconplish
this, I had to move the parsing of ANSI codes to where we read in the
content initially. To mitigate the additional memory penalty this would
incur, I switched to using ANSI tags that know how long the span of text
they should apply to is, but not the text itself.

This allowed me to preserve the ANSI SGR state (`style`) across lines.