ociruntime: no-op cleanup (#10031)
This is a preparation patch for #9978.

First, change to only collect network statistics when c.network is
non-nil.
In some test cases, network can be nil and cause a panic with the
existing
code.

Second, remove the redundant `--persistent_worker` flag from
TestPersistentWorker
as persistentworker.Start should be adding that flag automatically.