Fix target filtering in dense mode (#10037)
In dense mode, the targets tab is the default tab when no hash is set -
which breaks InvocationFilterComponent (which expects the hash to be
targets).

This PR fixes that by passing down the active tab to the filter
component (which is correct regardless of whether or not we're in dense
mode).

Fixes: https://github.com/buildbuddy-io/buildbuddy/issues/8829