Add a util for copying outgoing RPC metadata to incoming RPC metadata (#10035)
We have a few places where we simulate RPCs in tests by creating a new
context and copying the RPC metadata from a context that's passed in to
the new context. This PR just pulls that few lines out of the few places
we have it into a little utility function, mostly because I want to add
another one soon and thought four of them was too many.