[raft] return errors when partition.NumRanges is set to > 10 (#10081)
In order to support a large p.NumRanges, we want to batch the number of
ranges we initialize and also have partition descriptor where we can
store the state of the initialization so that we can recover from
failures. Before this happens, I don't want us accidentally put very
large number of statements in a transaction.