Switch passthrough github endpoints to `GetGitHubAppForAuthenticatedUser` (#10087)
These are the endpoints that are called by the code v2 / pull request
UIs.

`GetGitHubAppForOwner` requires the repo be installed / linked, but we
want to be able to edit code for any public repo.

Without this if you try to edit the bazelbuild repo in dev, you get
`NotFound: no GitHub app installation for "bazelbuild" was found`.