Fix line-wrapping issue that causes log lines to not be erased by the ANSI cursor (#10040)
This PR addressess the issue by setting the maximum columns of the
emulated ANSI window to be very large, but sets the initial allocated
length to a more moderate size.

Happy to accommodate other opinions regarding window sizes, these were
just some arbitrary values that seemed large enough and fixed the
problem when building buildbuddy.

As a precautionary measure, also sets a maximum number of lines to
retain so that the window size cannot balloon enormously if bazel sends
us logs with an unusually large cursor window.

This issue is caused by having a window size with too few columns for
the output, resulting in line wraps in the output. Then when bazel moves
the cursor up to erase the current progress, the cursor moves up one
fewer line than bazel expects for every line wrap that occurred. This
results in "leaving behind" progress lines. We can fix it by increasing
the column count of the window, but the column count is the actual
capacity of the screenline slices in go, so making this number big means
using a lot of memory that, for most lines, is completely unnecessary. I
fixed this in the library by adding a `defaultColumnCapacity` parameter,
which decouples the window size from the allocated capacity, allowing us
to accommodate a much larger line size without ballooning memory.