Add ClickHouse backup tool (#10066)
Adds a script to backup and restore ClickHouse data, as well as some
tooling updates to clickhouse_cluster + example configs to support
manual testing.
- Backups are taken at the database level (all tables are backed up from
a single database).
- Backups can be restored at the database level or individual table
level.
- A full backup is taken on the first of each month or if no recent
backups exist; otherwise an incremental backup is taken.
- GCS folder names determine the dates; we just check the most recent N
days to determine whether there is a recent backup.

I spent a little bit of time looking at
https://github.com/Altinity/clickhouse-backup as an alternative, but I
got the impression that it wasn't worth using it:
- That tool predates the ClickHouse-native `BACKUP` / `RESTORE` queries,
so it has a lot of extra functionality built in which is now supported
natively using those queries.
- The native `BACKUP` / `RESTORE` queries are supported by the tool but
they are not used by default, and require extra configuration to set up.
- It seems like incremental backups require using either the "watch"
mode (running the tool continuously) or writing some scripting around
individual backups. If we're having to write scripts around the tool
anyway, I figured we may as well just write our own little tool that
includes this logic, as well as issuing the `BACKUP` / `RESTORE`
queries.