WebUI: Ignore unset experimental_remote_cache_compression_threshold for Bazel 8 (#10131)
Starting in Bazel 8,
`--experimental_remote_cache_compression_threshold`'s default value
changes from 0 to 100. The existing warning recommendation is a false
positive for Bazel 8 workspaces. Check for Bazel 8 to not show the
recommendation. If a Bazel 8 workspace explicitly sets
`--experimental_remote_cache_compression_threshold` as well in Bazel 8
(even to 0), then the recommendation is not required anyways.