Fix firecracker not respecting numeric dockerUser (#10161)
This command works with OCI isolation type:

```
$ bb execute --exec_properties=dockerUser=1001 -- id
uid=1001 gid=0(root) groups=0(root)
```

As expected, this matches the behavior of `docker run`:

```
$ docker run --user=1001 gcr.io/flame-public/rbe-ubuntu22-04 id
uid=1001 gid=0(root) groups=0(root)
```

However on firecracker, it unexpectedly results in a uid lookup error:

```
$ bb execute --exec_properties=dockerUser=1001 --exec_properties=workload-isolation-type=firecracker --remote_header=x-buildbuddy-api-key=${BB_API_KEY?} -- id
failed to receive from stream: uid lookup failed: user: unknown userid 1001
```

This is unexpected because we mostly want firecracker to be a drop-in
replacement for docker/podman/OCI isolation types.

I ran into this trying to set a numeric `dockerUser` on a unit test,
which executes in a special image with some utils provisioned for the
test, but which doesn't have a named non-root user provisioned. Being
able to set a numeric user ID is useful because it allows running as
non-root without having to actually provision a user in the image (which
can be inconvenient or impractical)