Fix custom resource starvation and generalize queue skipping (#10109)
Context:
https://buildbuddy-corp.slack.com/archives/C06DBLTGC2V/p1755208425086889?thread_ts=1755207782.144469&cid=C06DBLTGC2V

Today, if a task is _only_ blocked on custom resources (e.g. there's
plenty of CPU and RAM but not enough GPU), then we skip that task in the
queue and check whether the next task can be scheduled. This logic was
mostly just intended to allow tasks which only need CPU and RAM
resources (e.g. compile actions) to skip ahead of actions which were
stuck waiting for custom resources to become available.

However, the current logic can easily result in starvation if the
executor is busy. Let's say we're in this state:
```
capacity:         cpu=8 ram=8G gpu=8G
used:             cpu=1 ram=1G gpu=1G
queue[0] (head):  cpu=1 ram=1G gpu=8G (exclusive GPU access)
queue[1]:         cpu=1 ram=1G gpu=2G
```

In this scenario, `queue[0]` is only blocked on GPU resources, so we
just skip it and check whether the next task can schedule. We then allow
`queue[1]` to run, because there is 7GB of schedulable memory remaining
(`capacity - used`), which is more than its 2GB request. If the executor
is busy executing lots of smaller gpu tasks like `queue[1]`, then we'll
never wind up dequeueing `queue[0]`.

As an even more extreme example, imagine that `used gpu=0.001GB` and
`queue[1]` needs `7.999GB` of gpu. Its task size is practically equal to
`queue[0]`, yet we're letting it skip ahead for no reason.

The fix is to implement "backfilling" which is a scheduling policy in
which we let tasks skip the queue but only if they don't delay the start
time of the tasks ahead of them. To provide this guarantee, we keep
track of "reserved" resources when looking through the queue for the
next task to dequeue. When we skip a task, we "reserve" the resources
for that task, effectively adding it to the `used` resource counts. So
in the above example, we would have reserved 9GB of GPU memory after
considering `queue[0]` in the queue, and then `queue[1]` would not be
allowed to schedule, since 9GB is greater than the capacity (8G).

While implementing this, I realized there isn't much of a reason to
restrict this logic to custom resources, so I made it generalize to CPU
and RAM as well, but note that the new backfilling logic still only
applies if custom resources are enabled.