Lower pool size for client connections that are direct (#10125)
When we don't go through a proxy or load balancer, we can handle
unlimited streams per client connection, so there's no reason to create
15 connections.

When connecting to localhost, we don't need to worry about the health of
the peer, so we can totally skip pooling.

Lowering the number of client connections lowers the number of
goroutines used by the targeted servers, and this seems to lower heap
size and GC duration.