Don't set the inode64 mount option for /dev/shm (#10189)
Context:
https://buildbuddy-corp.slack.com/archives/C091JA2J1RU/p1756840184665559?thread_ts=1756507672.036899&cid=C091JA2J1RU

The kernel will set this option automatically if `CONFIG_TMPFS_INODE64`
is set. Setting the option explicitly results in an "invalid argument"
error if the option is unsupported.

I could not find any references to "inode64" in the podman codebase, so
it seems podman does not set this option explicitly either.