[raft] fix a bug in add replica (#10191)
If in AddReplica, after the replica descriptor is added to the staging
replica, we failed to add it as a non-voter in raft. When AddReplica
tries to run again, it tries to reserve a new replica ID and add it to
staging replicas of the range descriptor. Even if the second time it
succeeded, we will have a stale staging replica in the descriptor. The
driver will try to resolve it by calling AddReplica again. But this
time, verifyRequest fails since there is a replica descriptor with the
requested NHID in the replica descriptor.