Support workspace reuse for fuse workspaces. (#10064)
Since the filesystem layout is changing without going through the
kernel, it's on us to tell the kernel to invalidate information that it
may have cached for directories/files that may have been affected.

When updating the layout of a filesystem in vfs_server, keep track of
which inodes and inode entries need to be invalidated. Pass the
invalidated entries to vfs_unix which is responsible for notifying the
kernel that the entries have changed.