Skip image registry requests in a few more cases (#10200)
- Skip registry ping (`/v2/`) if the image is cached. The fix is to just
move `createRemoteLayer()` to later in the `Compressed()` function
(after the cache check), since this seems to be what causes the ping
request.
- Skip HEAD request to resolve the image to a digest if we already know
the digest (i.e. the container-image string already contains a digest).
The HEAD request just gives us an image descriptor (digest, length,
content-type), but if we have a digest ref and we have the manifest
cached, then we have everything we need already and can build the
descriptor directly, without hitting the registry.

Work towards https://github.com/buildbuddy-io/buildbuddy/pull/10199