[raft] stop replica when the partition descriptor is marked soft deleted (#10212)
Stop the replica in zombie janitor when the partition descriptor is
marked soft
delete.

The data is not removed, so when the server is restarted, the shard will
be
opened again; and stopped again by zombie janitor.

Also reads the partition from meta range in zombie janitor, and skip the
range
IDs that are in the middle of partition set up.