split out core queueing+batching logic in atime_updater for reuse (#10195)
we can reuse this functionality to define async migration / copy logic
between two caches.

current tests are basically just a dupe of atime_updater, with the logic
pulled into the test. it's a reasonable set of things to exercise the
code, dunno how much we want to delete from one or the other's tests
(theoretically, for example, we could just test the op function in the
atime updater instead of testing the internals of queueing)