Copy cached container image artifacts before replaying actions (#10220)
- Before replaying, attempt to copy cached image manifests and blobs
from the replay source to the replay target. This is similar to how we
copy inputs etc. before replaying.
- In the case where all artifacts are found in the cache, the image
registry can be completely bypassed. This requires setting the
`container-registry-bypass` platform prop, which was implemented in an
earlier PR. This property can only be used if authenticated using a
server admin key with special capabilities.
- Add `source_client_identity_key` and `target_client_identity_key` and
attach the signed client identity headers to the respective outgoing
contexts, since a signed client identity is required in order to
interact with the image cache.

One limitation here is that the executors do not actively try to cache
image layers in the case where the layers are already present in the
local OCI disk image cache. I hit this during local development, but
only after manually clearing the local app cache. If this becomes a
major issue, we could maybe have the executors actively call FindMissing
to try and keep the cached image layers fresh in the cache.