Resolve some problems with ascii processing / rendering (#10216)
This fixes some problems with window size, namely that
`ui_actions_shown` is not the number of lines we necessarily need, since
sometimes ui actions, however briefly, may wrap. This PR assumes ui
actions will not wrap more than once, since we do have to set an upper
limit on window size and for performance reasons it is nice if that
limit is lower. If we're wrapping more than once, the window width is
probably too small for the log to be particularly readable anyway.

Additionally, it removes the second code path for non-curses output.
There's no real reason to support that separately, since the curses
solution works just fine for it.

Should resolve buildbuddy-io/buildbuddy-internal#5026