Fix chained incremental backups not working (#10236)
We're checking whether the previous day's backup exists by looking for
the DB metadata file, but this file doesn't seem to get created for
incremental backups. As a result, we're taking a full backup every other
day, instead of creating a longer chain of incremental backups where
each one is based on the previous day.

Check for `.backup` instead, which gets created for both full and
incremental backups.