Validate default_isolation_type flag (#10249)
Context:
https://buildbuddy-corp.slack.com/archives/C091JA2J1RU/p1757519819422909?thread_ts=1757469653.024489&cid=C091JA2J1RU

Based on the error message and the provided invocation link
(specifically the platform props shown for the failing executions), it
looks like the customer has `default_isolation_type: oci` configured,
but not `enable_oci: true`. This is a configuration error that we can
surface in a more obvious way.