[remote exec] resolve image tags to digests (behind flag) (#10259)
This change introduces the `--executor.resolve_image_digests` flag.
If enabled, container image names will be resolved to include a digest
during image warmup and when getting a runner bound to a task.

A customer has encountered HTTP 401 on the HEAD request that resolves
the image name to a digest. We have not. Using a flag here so we can
exercise this code in production without breaking downstream customers.