User list management UI.
This PR contains the basic crud operations for "user lists":
  - create
  - delete
  - rename
  - update members

Does not contain changes to allow permissions to be assigned to user
lists.

Note: in the UI "user lists" are called "Groups" since this is what they
are usually called, but let me know if you think that's too confusing.

This change depends on the backend changes in
https://github.com/buildbuddy-io/buildbuddy/pull/10315