Add ClickHouse schema for usage (#10182)
Add the `RawUsage` table schema and associated constants for use with
the low-cardinality string fields.

We still have some work to do before we start officially relying on this
data, but I figured it doesn't hurt to start defining the schema and
storing data in the DB so we can start making use of it for some
internal exploratory analytics.

Doc:
https://docs.google.com/document/d/1jv_gIMpbwIKvDrcLRRWEYA02Vkv5nvXp1i0Tdg0YKbA/edit?tab=t.0#heading=h.l4o7xrhqb6bt