Fail on startup if http port is busy (#10284)
This caused me some confusion during local development - the executor
will happily start up if something is listening on port 8888 already,
but it should probably fail. In my case I forgot I was running some
pprof web stuff listening on port 8888, and I was confused when I tried
to check /readyz on port 8888 (for the executor) and it gave me a 404.