Make role field in GroupRole optional. (#10267)
With the introduction of indirect memberships, a user could have more
than one role assigned. In those cases, we'll not populate the role at
all.

SCIM is the last remaining user of the role field and will only look at
direct memberships for user attributes in which case it can continue to
use the role field until everyone has been migrated to use uesr lists.