refactor: move server admin flag / auth logic to claims package, and simplify (#10286)
Context: I'd like to add a new field `ExperimentTargetingGroupID` to
`Claims`, which will be based on a gRPC header that only server admins
should be able to set. The current code structure makes this difficult
to implement - `ClaimsFromContext` does not currently get passed an
`env` or an `Authenticator`, but the authenticator is needed to know the
server admin group ID (using the `AdminGroupID()` method).

So, this PR shuffles things around so that `claims.go` has more direct
access to the server admin group ID. In particular, it makes the server
admin group ID a plain old flag, which makes things simpler overall.