[raft] in InitializeClusters log the error returned from matchesListenAddress (#10307)
When we start raft clusters from scratch, sometimes the machines
restarted
because matchesListenAddresses return no such host error. It can happen
when the
machine doesn't have info about other machines yet. When it returns
no-such-host
error, it indicates that it's not the bringup coordinator. This also
matches the
other callsite of this function.