Add default line length for emulated terminal for build logs (#10263)
Sometimes we do not receive the build event containing the terminal line
length before we are forced to begin logging, and we should have a sane
default for that. We were falling back to the max length before this PR,
which is 2000, but that resulted in too few max line lines, since that
is governed by max window size / max line length, so we were getting the
"leaving behind lines" behavior in this case.