Fix JVM sometimes not respecting cgroup2 memory limits (#10323)
In some of our newer executor deployments, the `cpuset` row is missing
from `/proc/cgroups`, which prevents the JVM from respecting container
memory limits. Work around this problem by faking out this file (which
should be safe, since it's already a fake file - it's a cgroup v1
specific file that has been kept in cgroup v2 to avoid breaking older
apps).

See comment on `getFakeProcCgroupsContent` for more detail.

Fun fact: the JVM was reporting that the `cpu` controller was disabled,
not the `cpuset` controller, which made things even more confusing - it
turned out this was yet another JVM bug that was also fixed
[here](https://github.com/openjdk/jdk/commit/2de71d04454b04ee887f7bd3e5decbfaa9ab8460)
(the `cg_controller_name` list was ordered incorrectly).