Fix flake due to healthchecker not shutting down (#10303)
Add `WaitForGracefulShutdown` to test env cleanup because not doing it
was causing flakes.


[Example](https://buildbuddy.buildbuddy.dev/invocation/82e4d37a-6801-43f8-b2c7-31425d4c27fa?target=%2F%2Fenterprise%2Fserver%2Fapi%3Aapi_test&targetStatus=6)

Basically, there is a race condition around setting flags because they
can be used during the healthchecker shutdown. If we wait for shutdown,
this race condition disappears.

However, this means we have to shut down the `priority_task_scheduler`
in cleanup for the `priority_task_scheduler` tests. Which also means we
have to complete all the scheduler's tasks. Initially, this caused the
task scheduler to hang, so this PR also switches the way we track how
many cancel tasks are active and how to wait for them to finish.