[cli] show help for `bb` and `bazel` commands (#10326)
Prior to this change, `bb help remote` (or any other BB CLI command)
would print an error message saying the `remote` command could not be
found. This is because we were always routing to `bb help <command>` to
`bazelisk`.

This change has `bb help <command>` determine whether `<command>` is a
BB CLI command. If so, it will print the help for that command and exit.
Otherwise, it will invoke `bazelisk help <command>`.

This change adds tests for all BB CLI commands currently in the source
tree and all Bazel commands present in `8.4.0`.

This change DOES ~not~ test `bb <command> -h` or `bb <command> --help`.