Fix tasks hanging if persistent worker crashes (#10351)
`exec.Cmd.Wait` never returns in the case where `cmd.Stdin` is not an
instance of `*os.File` and it's not receiving any input, even if the
underlying process has exited already.

This means if a persistent worker crashes after we've already sent the
persistent work request, we'll hang indefinitely due `exec.Cmd` being
stuck.