Fix bug in cachetools.UploadWriter (#10358)
We would return io.ErrShortWrite instead of status.AlreadyExists when
the resource already exists.

This is a bit clunky since byte_stream_server.go doesn't guarantee that
it will notify the client if the object already exists. So in
UploadWriter.Write, we do a best effort check so the caller can
short-circuit, but in Commit, we treat that as success for consistency.