Use references to external repo where possible (#10364)
Each label attribute specified in MODULE.bazel is modified as follows:
* if it can refer to an arbitrary repository, it is prefixed with the
go_deps-style repo name of the public repo, thus allowing this reference
to remain valid when it is loaded as a repo in the internal repo.
* if it can't because it is an attribute on an override, refer to the
current module instead. Such files will have to be mirrored.

This requires aligning the repo name of the public repo with the name
used for it by the internal repo.