User list role assignment. (#10374)
Groups are shown at the top of the member list (both assigned and
unassigned).
I expect the number of groups per organization to be small so having
them accessible directly in the member list is convenient.
<img width="1297" height="767" alt="image"
src="https://github.com/user-attachments/assets/514f94ae-b028-4305-926d-f28eb4a82d3b"
/>

Role assignment and removal works the same as it does for users (except
that removing role from a user list does not remove it from the UI)