Remove noisy warning from retry package (#10385)
In replay_action, this warning can wind up getting printed thousands of
times when pressing Ctrl+C, which makes it harder to go back and look at
log output.

I think this warning is not needed because an error is already being
returned to the caller, who can decide how best to handle the error.
It's also a context cancelled error, which is generally not that
interesting to log.