More fixes and improvements for replay_action (#10381)
* Fix a bug where setting `--n` (runs per action) to a value greater
than 1 would not work as expected due to action merging (several of the
replay attempts would be merged). The fix is to reupload and replay a
modified action which has `do_not_cache` set to true (if it isn't set
already), since this seems to be the only way to get the execution
server to disable action merging.
* Allow the tool to create a BES stream so that the results can be
queried from the UI, and (more importantly) the execution data can be
stored in ClickHouse for easier querying.
* As part of this change, I had to replace all occurrences of
`log.Fatal` with returning errors, so that the BES stream can be
properly closed on failure, instead of the process terminating (which
disconnects the invocation)
* Also as part of this change, `action_mnemonic` and `target_label` are
now sent up in the execution RequestMetadata. This is needed to populate
the respective ClickHouse columns, which are useful for grouping. This
is also needed for proper experiment targeting, as these things are
included in the experiment evaluation context.
* When replaying from a bundle, allow `--execution_id` to filter
executions from the bundle. Also, use more flexible parsing for
`--execution_id` (allow it to be either an upload string or download
string)
* Fix `Ctrl+C` being ignored when replaying to or from a bundle. The
same fix is also needed to properly terminate the BES stream if
applicable, even when not using a bundle.
* Fix the default `grpc_port` clashing with the default local app port
(change the default from 1985 -> 11985)
* Add a `slow_task_threshold` flag (defaults to 10m) to more easily
diagnose when executions are stuck. (When replaying lots of executions,
it's hard to know exactly what's stuck, especially if running with
`--app.log_level=warn`)